meson-tests: enable running with no hugepages
Various test environments don't always include hugepage support.  In order
to run the tests in these systems, expose an option that will pass the
no-huge flag to the EAL.  This could be auto-detected, but it also makes
sense to pass as a flag in order to test --no-huge support even in those
environments where hugepages are configured.

Signed-off-by: Aaron Conole <aconole@redhat.com>